refactor: TodoListView 수정 #10
